+++ test/noahtheduke/splint/netrunner_test.clj
@@ -32,6 +32,7 @@
     metrics/parameter-count 204
     naming/conventional-aliases 33
     naming/conversion-functions 28
+    performance/get-keyword 5
     naming/predicate 15
     performance/assoc-many 724
     style/apply-str 1
@@ -83,5 +84,5 @@
              :diagnostics
              (group-by :rule-name)
              (#(update-vals % count)))))
-    (expect 2361 (count (:diagnostics results)))
+    (expect 2366 (count (:diagnostics results)))
     (expect 222 (count (:checked-files results)))))